Fully implemented character input.
Implemented the function of forcing the timer to tick upon the first
click of a button, which sets all of the buttons state to false through
'boolButtonPressed', sets the current buttons state to true, then
performs the code previously written.

Also changed the characters used in the * button to the correct
characters.